+++ GrokkingCoding/src/com/fruitfactory/slidingwindow/wordconcatenation/WordConcatenationTests.java
@@ -1,7 +1,7 @@
 package com.fruitfactory.slidingwindow.wordconcatenation;
 
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
+import org.junit.Test;
+import static org.junit.Assert.*;
 import java.util.List;
 
 public class WordConcatenationTests {
@@ -11,8 +11,9 @@ public class WordConcatenationTests {
         WordConcatenation wc = new WordConcatenation();
         List<Integer> result = wc.findWordConcatenation("catfoxcat", new String[]{"cat", "fox"});
 
-        assertEquals(0,result.get(0));
-        assertEquals(3,result.get(1));
+        assertEquals(0,result.get(0).intValue());
+        assertEquals(3,result.get(1).intValue());
+
     }
 
 
@@ -21,7 +22,7 @@ public class WordConcatenationTests {
         WordConcatenation wc = new WordConcatenation();
         List<Integer> result = wc.findWordConcatenation("catcatfoxfox", new String[]{"cat", "fox"});
 
-        assertEquals(3,result.get(0));
+        assertEquals(3,result.get(0).intValue());
     }
 
     @Test
@@ -29,8 +30,8 @@ public class WordConcatenationTests {
         WordConcatenation wc = new WordConcatenation();
         List<Integer> result = wc.findWordConcatenation2("catfoxcat", new String[]{"cat", "fox"});
 
-        assertEquals(0,result.get(0));
-        assertEquals(3,result.get(1));
+        assertEquals(0,result.get(0).intValue());
+        assertEquals(3,result.get(1).intValue());
     }
 
 
@@ -39,6 +40,6 @@ public class WordConcatenationTests {
         WordConcatenation wc = new WordConcatenation();
         List<Integer> result = wc.findWordConcatenation2("catcatfoxfox", new String[]{"cat", "fox"});
 
-        assertEquals(3,result.get(0));
+        assertEquals(3,result.get(0).intValue());
     }
 }
